requirements.txt
python-telegram-bot==11.1.0
requests
python-whois==0.7.1